sample-app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
PyJWT>=2.8.0
msal>=1.28.0
azure-storage-blob>=12.19.0
redis>=5.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
